+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug #37: Move `Message` enum to `Helper\Exception` namespace and clean up unused cases (@terabytesoftw)
 - Bug #38: Update method calls to use `getAttributes()` method for consistency in `BaseBlock`, `BaseInline`, and `BaseVoid` classes (@terabytesoftw)
 - Bug #39: Rename exception test methods for clarity and consistency (@terabytesoftw)
+- Bug #40: Consolidate attribute handling in `setAttribute()` method in `HasAttributes` class and remove redundant code in related classes (@terabytesoftw)
 
 ## 0.3.1 December 26, 2025